pyproject.toml
[project]
name = "leekwars-agent"
version = "0.1.0"
description = "LeekWars automation and RL agent"
authors = [
    {name = "pln"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "playwright (>=1.57.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "gymnasium (>=1.2.3,<2.0.0)",
    "stable-baselines3 (>=2.7.1,<3.0.0)",
    "sb3-contrib (>=2.7.1,<3.0.0)",
    "tensorboard (>=2.20.0,<3.0.0)",
    "py4j (>=0.10.9.9,<0.11.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "websocket-client (>=1.9.0,<2.0.0)"
]


[project.scripts]
leek = "leekwars_agent.cli.main:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)"
]
